Refactored AddressManager to be used by web3 and google addresses
